requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
plotly>=5.18.0
numpy>=1.26.0
python-dateutil>=2.9.0
pydantic>=2.0.0
